fix: extend CAST to structured types
fixes: https://github.com/confluentinc/ksql/issues/6584

Extend CAST to the structured types `ARRAY`, `MAP` and `STRUCT`.
