+++ ksqldb-execution/src/test/java/io/confluent/ksql/execution/codegen/SqlToJavaVisitorTest.java
@@ -36,6 +36,7 @@ import static org.mockito.Mockito.when;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
+import io.confluent.ksql.execution.codegen.helpers.CastEvaluator;
 import io.confluent.ksql.execution.expression.tree.ArithmeticBinaryExpression;
 import io.confluent.ksql.execution.expression.tree.ArithmeticUnaryExpression;
 import io.confluent.ksql.execution.expression.tree.ArithmeticUnaryExpression.Sign;
@@ -60,7 +61,6 @@ import io.confluent.ksql.execution.expression.tree.StringLiteral;
 import io.confluent.ksql.execution.expression.tree.SubscriptExpression;
 import io.confluent.ksql.execution.expression.tree.TimeLiteral;
 import io.confluent.ksql.execution.expression.tree.TimestampLiteral;
-import io.confluent.ksql.execution.expression.tree.Type;
 import io.confluent.ksql.execution.expression.tree.UnqualifiedColumnReferenceExp;
 import io.confluent.ksql.execution.expression.tree.WhenClause;
 import io.confluent.ksql.function.FunctionRegistry;
@@ -72,18 +72,15 @@ import io.confluent.ksql.function.types.ParamTypes;
 import io.confluent.ksql.function.udf.UdfMetadata;
 import io.confluent.ksql.name.ColumnName;
 import io.confluent.ksql.name.FunctionName;
-import io.confluent.ksql.name.SourceName;
 import io.confluent.ksql.schema.Operator;
-import io.confluent.ksql.schema.ksql.types.SqlDecimal;
 import io.confluent.ksql.schema.ksql.types.SqlPrimitiveType;
 import io.confluent.ksql.schema.ksql.types.SqlTypes;
 import io.confluent.ksql.util.KsqlConfig;
+import io.confluent.ksql.util.Pair;
 import java.math.BigDecimal;
 import java.util.Collections;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.stream.Collectors;
-
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -92,8 +89,6 @@ import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
 
 public class SqlToJavaVisitorTest {
-
-  private static final SourceName TEST1 = SourceName.of("TEST1");
 
   @Mock
   private FunctionRegistry functionRegistry;
@@ -222,54 +217,16 @@ public class SqlToJavaVisitorTest {
         COL0,
         new io.confluent.ksql.execution.expression.tree.Type(SqlPrimitiveType.of("INTEGER"))
     );
-    final Expression castDoubleBigint = new Cast(
-        COL3,
-        new io.confluent.ksql.execution.expression.tree.Type(SqlPrimitiveType.of("BIGINT"))
-    );
-    final Expression castDoubleString = new Cast(
-        COL3,
-        new io.confluent.ksql.execution.expression.tree.Type(SqlPrimitiveType.of("VARCHAR"))
-    );
+
+    // When:
+    final String actual = sqlToJavaVisitor.process(castBigintInteger);
 
     // Then:
-    assertThat(
-        sqlToJavaVisitor.process(castBigintInteger),
-        equalTo("(new Long(COL0).intValue())")
-    );
-    assertThat(
-        sqlToJavaVisitor.process(castDoubleBigint),
-        equalTo("(new Double(COL3).longValue())")
-    );
-    assertThat(
-        sqlToJavaVisitor.process(castDoubleString),
-        equalTo("Objects.toString(COL3, null)")
-    );
-  }
+    final String expected = CastEvaluator
+        .eval(Pair.of("COL0", SqlTypes.BIGINT), SqlTypes.INTEGER, ksqlConfig)
+        .getLeft();
 
-  @Test
-  public void shouldUseStringValueOfIfConfigSet() {
-    // Given:
-    final Expression castDoubleString = new Cast(
-        COL3,
-        new io.confluent.ksql.execution.expression.tree.Type(SqlPrimitiveType.of("VARCHAR"))
-    );
-    ksqlConfig = new KsqlConfig(Collections.singletonMap(KsqlConfig.KSQL_STRING_CASE_CONFIG_TOGGLE, false));
-    final AtomicInteger funCounter = new AtomicInteger();
-    final AtomicInteger structCounter = new AtomicInteger();
-    sqlToJavaVisitor = new SqlToJavaVisitor(
-        SCHEMA,
-        functionRegistry,
-        ref -> ref.text().replace(".", "_"),
-        name -> name.text() + "_" + funCounter.getAndIncrement(),
-        struct -> "schema" + structCounter.getAndIncrement(),
-        ksqlConfig
-    );
-    
-    // Then:
-    assertThat(
-        sqlToJavaVisitor.process(castDoubleString),
-        equalTo("String.valueOf(COL3)")
-    );
+    assertThat(actual, is(expected));
   }
 
   @Test
@@ -834,96 +791,6 @@ public class SqlToJavaVisitorTest {
 
     // Then:
     assertThat(java, is("(COL8.plus(new MathContext(2, RoundingMode.UNNECESSARY)))"));
-  }
-
-  @Test
-  public void shouldGenerateCorrectCodeForDecimalCast() {
-    // Given:
-    final Cast cast = new Cast(
-        new UnqualifiedColumnReferenceExp(ColumnName.of("COL3")),
-        new Type(SqlDecimal.of(2, 1))
-    );
-
-    // When:
-    final String java = sqlToJavaVisitor.process(cast);
-
-    // Then:
-    assertThat(java, is("(DecimalUtil.cast(COL3, 2, 1))"));
-  }
-
-  @Test
-  public void shouldGenerateCorrectCodeForDecimalCastNoOp() {
-    // Given:
-    final Cast cast = new Cast(
-        new UnqualifiedColumnReferenceExp(ColumnName.of("COL8")),
-        new Type(SqlDecimal.of(2, 1))
-    );
-
-    // When:
-    final String java = sqlToJavaVisitor.process(cast);
-
-    // Then:
-    assertThat(java, is("COL8"));
-  }
-
-  @Test
-  public void shouldGenerateCorrectCodeForDecimalToIntCast() {
-    // Given:
-    final Cast cast = new Cast(
-        new UnqualifiedColumnReferenceExp(ColumnName.of("COL8")),
-        new Type(SqlTypes.INTEGER)
-    );
-
-    // When:
-    final String java = sqlToJavaVisitor.process(cast);
-
-    // Then:
-    assertThat(java, is("((COL8).intValue())"));
-  }
-
-  @Test
-  public void shouldGenerateCorrectCodeForDecimalToLongCast() {
-    // Given:
-    final Cast cast = new Cast(
-        new UnqualifiedColumnReferenceExp(ColumnName.of("COL8")),
-        new Type(SqlTypes.BIGINT)
-    );
-
-    // When:
-    final String java = sqlToJavaVisitor.process(cast);
-
-    // Then:
-    assertThat(java, is("((COL8).longValue())"));
-  }
-
-  @Test
-  public void shouldGenerateCorrectCodeForDecimalToDoubleCast() {
-    // Given:
-    final Cast cast = new Cast(
-        new UnqualifiedColumnReferenceExp(ColumnName.of("COL8")),
-        new Type(SqlTypes.DOUBLE)
-    );
-
-    // When:
-    final String java = sqlToJavaVisitor.process(cast);
-
-    // Then:
-    assertThat(java, is("((COL8).doubleValue())"));
-  }
-
-  @Test
-  public void shouldGenerateCorrectCodeForDecimalToStringCast() {
-    // Given:
-    final Cast cast = new Cast(
-        new UnqualifiedColumnReferenceExp(ColumnName.of("COL8")),
-        new Type(SqlTypes.STRING)
-    );
-
-    // When:
-    final String java = sqlToJavaVisitor.process(cast);
-
-    // Then:
-    assertThat(java, is("COL8.toPlainString()"));
   }
 
   @Test
